readpassphrase_owned uses full capacity
Only using up to `buf.len()` was a mistake. `readpassphrase_owned` will
now use the buffer’s capacity, not its length, and will truncate the
buffer to length 0 on error.

This necessitated calling through to the `ffi` API to avoid invoking
Rust undefined behavior by constructing a maybe-uninitialized slice. It
also was practically speaking much easier to write in terms of the added
but not exposed `readpassphrase_mut` function, which allows easier error
handling at the cost of confusing semantics (when does `buf` still have
its backing capacity or not?)

Also added a password retry loop example using `readpassphrase_owned`.
This example was failing on the old `readpassphrase_owned` (since the
next call would only read up to the length of the last password.)