Documentation cleanup, new APIs
This changes `explicit_bzero` to take a `String` by value rather than a
`Vec` by `&mut` to make it more ergonomic with `getpass`.
`explicit_bzzero_vec` has been introduced with the old interface, for
vectors used by `readpassphrase`. (I decided against a version that took
`&mut [u8]` since it would be too easy to misuse it with retry loops.)

Removed the mention of OpenBSD from the readme since it might give the
impression that this crate was BSD-only or something; I retained the man
link though.

Also removed usage instructions from the readme, locating them instead
on [docs.rs](https://docs.rs) and in the doc comments in `src/lib.rs`.

Added `getpass` to the keywords for this library.