Refactor, further documentation cleanup
I decided to go a different direction, dropping `explicit_bzero` in
favor of a minimal in-crate `zeroize::Zeroize` replacement that may be
shimmed out for the real `zeroize` by turning on that feature flag. This
approach seems more flexible, allowing the same `zeroize` call to be
used on strings, vectors, or slices.