Release v0.6.0 (#5)
This change includes substantial documentation improvements and some new
APIs and cleanup work.

# Breaking changes
* The `zeroize` feature is no longer enabled by default. If it is
  disbled, this crate includes its own minimal implementation of
  `zeroize::Zeroize` that can zero the specific types relevant to this
  library (i.e., `Vec<u8>`, `String`, and `[u8]`.) This is exposed as
  `mod zeroize`; now, the `zeroize` feature simply swaps this out for
  [`zeroize::Zeroize`][0].
* I removed the CStr error fork since `FromBytesUntilNulError` cannot
  happen as long as `readpassphrase(3)` keeps its contract, which we
  already assume elsewhere. Now the user-facing `Error` is much simpler:
  either `readpassphrase(3)` itself encountered an IO error or we
  encountered non–UTF-8 data.
* I changed `readpassphrase_owned` to return a new `OnwedError` type
  that includes the buffer on it and zeroizes that buffer on drop.
  `Error` now implements `From<OwnedError>`, making this easy to use
  with the `?` operator.
* The `explicit_bzero` function introduced before has been deprecated in
  favor of using either this new trait or `zeroize` itself, and
  generally the crate now encourages using `zeroize` a bit more.

# README cleanup
I removed the mention of OpenBSD from the readme since it might give the
impression that this crate was BSD-only or something. The man link to
openbsd.org remains.

I also removed usage instructions from the readme, locating them instead
on [docs.rs](https://docs.rs) by way of the doc comment in `src/lib.rs`.

I added `getpass` to the keywords for this library.

# Optional `cc` feature
Now it is possible to disable `cc`, which stops the C `readpassphrase`
code from being compiled. This allows either foreign C builds (e.g. via
Bazel) or use of the system libc’s built-in `readpassphrase` on
platforms that have it.

# docs.rs build tweaks
The build for <https://docs.rs> now turns on the `zeroize` feature in
the hopes that this will allow the examples to compile, but specifically
includes the minimal `zeroize` shim rather than the re-export of the
`zeroize` crate module.

Alongside this, the examples now use `zeroize` again.

[0]: https://crates.io/crates/zeroize