Support back to rust 1.72.0
## Motivation
As a library, we’d like to support compiler versions going as far back
as we can.

## What changed
Moves back to edition 2021 and changes the code so that it can compile
going back to 1.72.0, which was released a day shy of two years ago. The
only changes needed were removing `unsafe` from the `extern "C"` block
and implementing `Error` from `std::error` instead of `core::error`. (We
need `std::{io, str::Utf8Error}` pretty deeply, so a `no_std` feature
does not seem worth it at the moment.)