Always do a dry-run publish, stop attesting
The crates.io Trusted Publishing setup seems to handle build provenance
already, rendering this somewhat unorthodox use of build attestations
apparently superfluous; crates.io itself is now the place that attests
to the fact that a given crate hash came from a given workflow run.

At the same time, we’d like to get feedback from a `cargo publish` on
every push to main, so we always run the dry-run, and reserve the auth
and publish steps for tagged releases.