Be more defensive in `readpassphrase_mut`
We no longer need it to be the case that `res` is a pointer to the start
of `buf`, nor that the `readpassphrase` implementation produces a null-
terminated string, for Rust’s safety guarantees to be met. We do still
need there to be a null byte within `isize::MAX` of the start of `buf`;
one way to ensure that would be to write a null byte to the end of
`buf`’s capacity after we come back from `readpassphrase`, but I stopped
short of that.