Change OwnedError API
Adds `error` to fetch the `Error` and `into_bytes` to retrieve the
buffer while consuming ownership. Marks `take` deprecated and restores
it to the 0.9.x interface of `unwrap_or_default` since we won’t be
keeping it around much longer.

Reorders some impls so all the `OwnedError` impls are next to each other
and all the `Error` impls are next to each other.