Initialize `buf` in `readpassphrase_into`
This prevents us from invoking Rust undefined behavior in cases where a
`ffi::readpassphrase` implementation fails to initialize `buf`’s memory
but returns non-null. The performance cost is below the noise floor for
a routine that is generally reading a password from a TTY. To be extra
careful, we now only use (and initialize) up to 4KiB of `buf`’s spare
capacity.