Preserve `buf`’s contents on UTF-8 error
Changes `readpassphrase_into` so that it preserves the contents of `buf`
if the UTF-8 conversion failed. Requires calling `String::from_utf8`
instead of `CStr::to_str` to keep track of the bytes used.

We could have used `CStr::from_bytes_until_nul` again, but we’d need to
find the length of the byte slice, and here we run into trouble:
`CStr::count_bytes` is not available until Rust 1.79.0, and while
`CStr::to_bytes` is documented to be a constant-time cast right now,
there seems to be an intention to change that. So instead we do our own
scan for a zero byte.