Expand build script, add external feature
The build script will now exit nonzero if it fails to find a
readpassphrase implementation (unless the `external` feature is
enabled.)

This should hopefully make it a bit clearer when things go wrong, e.g.
due to libbsd not being installed but the `libbsd-static` feature being
enabled. It also establishes somewhat of a priority order among the
different places where the function can be found, with the vendored code
coming last.