Release v1.0.0 (#14)
This is the first stable release of this crate. From this point forward,
changes will be noted in a changelog, and any breaking changes will
result in a major version increase.

# Breaking Changes
* `readpassphrase_into` now only uses its input buffer’s spare
  capacity up to 4KiB; previously it always used the whole allocation.
* The `vendored` feature is gone; it has been split up into
  `linux-vendored` and `windows-vendored` features, both enabled by
  default.
* The `libbsd` feature has been renamed to `libbsd-static`.
* The deprecated names `readpassphrase_owned` and `OwnedError` have been
  dropped (in favor of `readpassphrase_into` and `IntoError`.)
* The crate build script will now fail if a readpassphrase
  implementation was not found, unless the `external` feature is
  enabled.

# New Features
* `Flags` now derives a number of common traits, namely `Debug`, `Copy`,
  `Clone`, `PartialEq`, `Eq`, and `Hash`.
* An `external` feature has been added to explicitly signal the intent
  to supply a `readpassphrase` implementation separately from the
  crate’s own build script.

# Other Changes
* `readpassphrase_into` and `getpass` now fully initialize their
  buffers before calling the C `readpassphrase` function, relaxing the
  requirements on the C function for Rust undefined behavior to not be
  invoked.
* If the `debug_assertions` `cfg` directive is enabled, then this
  crate’s functions will now panic in the event that the C
  `readpassphrase` implementation returns non-null but does not
  initialize its buffer to a byte sequence containing a zero byte. (If
  `debug_assertions` is off, then they will either panic or not
  depending on what the buffer’s well-defined prior state was.)
* The surface of unsafe code in this crate is now much reduced,
  consisting solely of the calls to the FFI `readpassphrase` function
  and a couple other trivially safe cases.
* The documentation has been improved, including describing the
  security requirements of `IntoError::into_bytes` and mentioning the
  previously-undocumented `readpassphrase_into` behavior on non-UTF8
  byte sequences (i.e., preserving the original bytes in the returned
  buffer.)

## License
This crate previously inherited the ISC license (at times mislabeled
by me as 1-clause BSD) from the BSD `readpassphrase` source code that
was previously vendored in-crate. Going forward, the crate itself is
made available under the user’s choice of either the Apache 2.0 or
MIT license (or the existing ISC license); the vendored code is now
distributed as an optional separate ISC-licensed crate.

This relicensing is not intended to be much (if any) of a change in
actual capabilities; the ISC license is a very permissive open-source
license, and if you are using a computer today you are almost certainly
using ISC-licensed software. My reason for doing this has more
to do with reducing the complexity of the legal component of the
decision-making process involved in using this crate. A lot of Rust
crates dual-license MIT or Apache; a lot of legal departments have
immediate green lights for one or the other of these licenses; the ISC
license seems less familiar in Rust-land and so might entail longer
conversations with legal than would be desired.